Use tippy.js to render tooltips
Tippy is much more flexible and lightweight, so this makes the script
run much more quickly.

Also removes the remote code execution that was previously present. This
was for the copyright stuff that the old API required, but it seems that
this is no longer necessary with the new API