Fix Chrome error from not responding to message
Every time a page was loaded with the extension enabled, an error would
be logged to the background worker that there was a port closing before
the script could respond. This was because the content script wasn't
responding to the background request in the listener. Adding an empty
sendResponse call fixed the issue